--- tutor/templates/hooks/cms/importdemocourse
@@ -1,8 +1,0 @@
-echo "Loading settings $DJANGO_SETTINGS_MODULE"
-
-# Import demo course
-git clone https://github.com/openedx/edx-demo-course --branch {{ OPENEDX_COMMON_VERSION }} --depth 1 ../edx-demo-course
-python ./manage.py cms import ../data ../edx-demo-course
-
-# Re-index courses
-./manage.py cms reindex_course --all --setup
